vfio/tio: Bind TDI to SNP VM
Running trusted IO requires binding a passed through PCI function
to a VM. The PCIe's TDISP protocol describes the state machine
which is implemented in the trusted agent (TSM). A part of setup is binding
a TDI (a descriptor of the TDISP state of a secure PCI function in
the physical adapted).

Add TDI_BIND call to VFIO device upon a request from the confidential VM.
The confidential VM makes such call from the device probing function so
such driver can enable the device first, receive the TDI interface report
and validate the device resources in the TSM.

This could potentially be done at the machine reset or on the device's
enablement (MMIO or MASTER in the command register), however TDISP
prohibits changes to MMIO and MASTER bits in the command register after
the binding step (TBD).

x-pre-bind is removed from now as not supported.

Tested with a SRIOV VF as:
echo -e device_add vfio-pci,host=e1:04.0,bus=r0,id=v0,iommufd=i0 | nc -q 0  -U ./qemu.mon.q.tvm

Signed-off-by: Alexey Kardashevskiy <aik@amd.com>